// src/app/modules/product/product.interface.ts
import { Types } from 'mongoose';

export interface IProduct {
  name: string;
  description: string;
  straight_up?: string; 
  lowdown: string[]; 
  regularPrice: number;
  salePrice: number;
  thumbnail: string;
  images: string[];
  categoryID: Types.ObjectId;
  stock: number;
  isFeatured: boolean;
  isBestseller: boolean;
  isNew: boolean;
}